move tests to new file so they run on ci
